Updated for Second Submission of project
Additional details added to the Write-Up for the project outlining various iterations during the creation of the poi_id.py file. 
The poi_id.py file was changed to change the features used, the way train/test data is created, and to test scaling of the data
Since the poi_id.py file creates the 3 pkl files these are updated as well.